Rename module name
