+++ client/blocks/plugins-scheduled-updates/hooks/use-is-eligible-for-feature.ts
