+++ client/blocks/plugins-scheduled-updates/hooks/use-prepare-plugins-tooltip-info.tsx
