+++ client/blocks/plugins-scheduled-updates/hooks/use-site-has-eligible-plugins.ts
