+++ client/blocks/plugins-scheduled-updates/schedule-list-cards.tsx
@@ -1,14 +1,14 @@
 import { Button, DropdownMenu, Tooltip } from '@wordpress/components';
 import { Icon, info } from '@wordpress/icons';
 import { useTranslate } from 'i18n-calypso';
-import { usePreparePluginsTooltipInfo } from 'calypso/blocks/plugins-update-manager/hooks/use-prepare-plugins-tooltip-info';
-import { useSiteDateTimeFormat } from 'calypso/blocks/plugins-update-manager/hooks/use-site-date-time-format';
-import { ellipsis } from 'calypso/blocks/plugins-update-manager/icons';
 import { useUpdateScheduleQuery } from 'calypso/data/plugins/use-update-schedules-query';
 import { Badge } from './badge';
 import { useIsEligibleForFeature } from './hooks/use-is-eligible-for-feature';
+import { usePreparePluginsTooltipInfo } from './hooks/use-prepare-plugins-tooltip-info';
 import { usePrepareScheduleName } from './hooks/use-prepare-schedule-name';
+import { useSiteDateTimeFormat } from './hooks/use-site-date-time-format';
 import { useSiteSlug } from './hooks/use-site-slug';
+import { ellipsis } from './icons';
 
 interface Props {
 	onEditClick: ( id: string ) => void;
@@ -74,9 +74,7 @@ export const ScheduleListCards = ( props: Props ) => {
 									{ prepareDateTime( schedule.last_run_timestamp ) }
 								</Button>
 							) }
-							{ schedule.last_run_status && (
-								<Badge type={ schedule.last_run_status } />
-							) }
+							{ schedule.last_run_status && <Badge type={ schedule.last_run_status } /> }
 						</span>
 					</div>
